requirements.txt
asgiref==3.8.1
crispy-tailwind==1.0.3
Django==5.2
django-allauth==65.7.0
django-crispy-forms==2.4
gunicorn==23.0.0
packaging==25.0
pillow==11.2.1
sqlparse==0.5.3
tzdata==2025.2
